backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0

# ML / NLP
torch
transformers==4.40.0
scikit-learn==1.4.2
lightgbm==4.3.0
numpy==1.26.4
scipy==1.13.0

# XAI
shap==0.45.0
grad-cam==1.5.0

# Image / video
Pillow==10.3.0
opencv-python-headless==4.9.0.80
torchvision

# Deepfake
# EfficientNet via torchvision

# Email / IMAP
imapclient==3.0.1

# Security / validation
dnspython==2.6.1
tldextract==5.1.2
python-whois==0.9.4

# Logging / utilities
loguru==0.7.2
aiofiles==23.2.1
slowapi==0.1.9

# OpenAI
openai==1.23.2
